% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG.R
\name{calculate_RPKM}
\alias{calculate_RPKM}
\title{calculate the RPKM for the input matrix if implementable}
\usage{
calculate_RPKM(readcount)
}
\arguments{
\item{readcount}{input: data frame/list object}
}
\description{
calculate the RPKM for the input matrix if implementable
}
