% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG.R
\name{check_exonlength}
\alias{check_exonlength}
\title{check whether exon length attribute exists in the input file}
\usage{
check_exonlength(readcount)
}
\arguments{
\item{readcount}{input: data frame/list object}
}
\description{
check whether exon length attribute exists in the input file
}
