% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG.R
\name{check_genecovered_quality}
\alias{check_genecovered_quality}
\title{data quality control: check the number of genes being covered}
\usage{
check_genecovered_quality(readcount)
}
\description{
data quality control: check the number of genes being covered
}
