% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG.R
\name{check_sample_distance}
\alias{check_sample_distance}
\title{Generalized sample distance using the normalized dds data}
\usage{
check_sample_distance(normalized_dds)
}
\arguments{
\item{normalized_dds}{normalized dds data generated by normalize_dataset()}
}
\description{
Generalized sample distance using the normalized dds data
}
