% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment_visualization.R
\name{develop_Gene_Network}
\alias{develop_Gene_Network}
\title{Gene-Concept Network Plot}
\usage{
develop_Gene_Network(
  edo,
  geneList,
  OrgDb = "org.Hs.eg.db",
  keyType = "ENTREZID",
  category_size = "pvalue"
)
}
\arguments{
\item{edo}{edoobject}

\item{geneList}{genelist object}

\item{OrgDb}{defaulted}

\item{keyType}{the key you have}

\item{category_size}{category_size}
}
\description{
visualize the potentially biological complexities in which a gene may belong to multiple annotation categories and provide information of numeric changes if available
convert gene ID to Symbol,to notice the default OrgDb and keytype we use is 'org.Hs.eg.db'(human) and
}
