% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment_visualization.R
\name{enrichmap}
\alias{enrichmap}
\title{Function to plot edo output as a network}
\usage{
enrichmap(edo, cex_category = 1.1)
}
\arguments{
\item{edo}{large enrichResult}

\item{cex_category}{The cex_category parameter can be used to resize nodes}
}
\description{
Function to plot edo output as a network
}
