% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment_visualization.R
\name{filter_genelist}
\alias{filter_genelist}
\title{filter_genelist}
\usage{
filter_genelist(geneList, standard_fc)
}
\arguments{
\item{geneList}{should be a list of gene foldchange level or with names of geneID}

\item{standard_fc}{the foldchange threshold for filtering}
}
\value{
edo object
}
\description{
This function select interested genes with abolute value greater than assigned threshold and convert to large enrichResult. l
}
