% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG.R
\name{load_data_for_DESeq2}
\alias{load_data_for_DESeq2}
\title{load data into DESeq2 object}
\usage{
load_data_for_DESeq2(file, condition_vector, type_vector)
}
\arguments{
\item{file}{the file/dataframe you want to use}

\item{condition_vector}{the condition of the samples}

\item{type_vector}{the type/design of samples}
}
\description{
load data into DESeq2 object
}
