% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG.R
\name{select_DEG}
\alias{select_DEG}
\title{differential gene expression and return their Entriz ID/SYMBOL for further gene enrichment analysis}
\usage{
select_DEG(dds, filter_thresh = 0, log2_fc = 2, padj = 0.05)
}
\description{
differential gene expression and return their Entriz ID/SYMBOL for further gene enrichment analysis
}
