% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment_visualization.R
\name{show_dotplot}
\alias{show_dotplot}
\title{Dotplot}
\usage{
show_dotplot(edo, showCategory_num)
}
\arguments{
\item{edo}{large enrichResult}

\item{showCategory_num}{specify the number of terms (most significant) or selected terms to display}
}
\description{
method to visualize enriched terms via dot plot
}
