% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG.R
\name{check_totalcov_quality}
\alias{check_totalcov_quality}
\title{data quality control: check the total coverage}
\usage{
check_totalcov_quality(readcount)
}
\arguments{
\item{readcount}{input: data frame/list object}
}
\description{
data quality control: check the total coverage
}
