% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment_visualization.R
\name{develop_Gene_Network}
\alias{develop_Gene_Network}
\title{Gene-Concept Network Plot
convert gene ID to Symbol,to notice the default OrgDb and keytype we use is 'org.Hs.eg.db'(human) and}
\usage{
develop_Gene_Network(
  edo,
  geneList,
  GeneNetp_name,
  OrgDb = "org.Hs.eg.db",
  keyType = "ENTREZID",
  category_size = "pvalue"
)
}
\arguments{
\item{edo}{edoobject}

\item{geneList}{genelist object}

\item{GeneNetp_name}{image_name}

\item{category_size, keyType, OrgDb}{defaulted}
}
\description{
Gene-Concept Network Plot
convert gene ID to Symbol,to notice the default OrgDb and keytype we use is 'org.Hs.eg.db'(human) and
}
