% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gene_enrichment_visualization.R
\name{filter_genelist}
\alias{filter_genelist}
\title{filter_genelist}
\usage{
filter_genelist(geneList, standard_fc)
}
\value{
edo object
This function select interested genes with abolute value greater than assigned threshold and convert to large enrichResult
}
\description{
filter_genelist
}
