% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_DEG.R
\name{select_DEG}
\alias{select_DEG}
\title{select differential gene expression}
\usage{
select_DEG(dds, filter_thresh = 0, log2_fc = 0.58, padjust = 0.05)
}
\arguments{
\item{dds}{the dds object}

\item{filter_thresh}{the threshold to remove genes with expression level lower than this}

\item{log2_fc}{log2foldchange to filter the significant genes}

\item{padjust}{adjusted p value to filter the significant genes}
}
\description{
select differential gene expression
}
